backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
uagents>=0.22.0
requests>=2.32.0
pydantic>=2.0.0
python-dotenv>=1.0.1
openai>=1.40.0
google-generativeai>=0.8.3
